backend/requirements.txt
fastapi==0.116.1
uvicorn==0.30.6
python-multipart==0.0.20
python-docx==1.2.0
PyMuPDF==1.24.9
openai==1.50.2
python-dotenv==1.0.1
